% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_chisq}
\alias{hex_chisq}
\title{Chi-squared test for HWE using Wang's approach.}
\usage{
hex_chisq(yww, nind, niter = 8, alpha = 0, method = c("correct", "incorrect"))
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{nind}{The number of individuals in the sample.}

\item{niter}{The number of EM iterations.}

\item{alpha}{The double reduction rate (assumed known).}

\item{method}{Should we calculate the chi-squared statistic the correct way
(\code{"correct"}) or the way that Wang et al calculated it
(\code{"incorrect"})? This also affects the degrees of freedom.}
}
\description{
Chi-squared test for HWE using Wang's approach.
}
\examples{
nvec <- c(29, 21, 17, 10, 10, 10, 23)
nind <- sum(nvec)
yww <- nvec / nind
hex_chisq(yww = yww, nind = nind, method = "correct")

## reproduce 6.602 value from page 5 of Wang et al (2022)
hex_chisq(yww = yww, nind = 120, method = "incorrect")

}
\references{
\itemize{
\item{J. Wang, L. Feng, S. Mu, A. Dong, J. Gan, Z. Wen, J. Meng, M. Li, R. Wu, and L. Sun. Asymptotic tests for Hardy-Weinberg equilibrium in hexaploids. \emph{Horticulture Research}, 9, 05 2022. \doi{10.1093/hr/uhac104}.}
}
}
\author{
Jing Wang and David Gerard
}
