% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_drtest}
\alias{hex_drtest}
\title{Test for HWE}
\usage{
hex_drtest(yww, nind, niter = 8)
}
\description{
Their code is super weird, so I try to reproduce it using simpler code.
}
\examples{
nvec <- c(29, 21, 17, 10, 10, 10, 23)
nind <- sum(nvec)
yww <- nvec / nind

## this should be 5.922
hex_drtest(yww = yww, nind = nind, niter = 8)

}
\references{
\itemize{
\item{J. Wang, L. Feng, S. Mu, A. Dong, J. Gan, Z. Wen, J. Meng, M. Li, R. Wu, and L. Sun. Asymptotic tests for Hardy-Weinberg equilibrium in hexaploids. \emph{Horticulture Research}, 9, 05 2022. \doi{10.1093/hr/uhac104}.}
}
}
\author{
David Gerard
}
