% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_em}
\alias{hex_em}
\title{EM algorithm to estimate gamete frequencies}
\usage{
hex_em(yww, niter = 30)
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{niter}{The number of EM iterations.}
}
\value{
The estimated gamete frequencies

A list of length 2.
\itemize{
\item{\code{p}: The estimated gamete frequencies.}
\item{\code{q}: The estimated genotype frequencies.}
}
}
\description{
This code was really really weird. I had to edit it slightly to make it
generalizable beyond the specific data of (29, 21, 17, 10, 10, 10, 23).
}
\examples{
yww <- c(29, 21, 17, 10, 10, 10, 23) / 120
hex_em(yww = yww)

}
\author{
Jing Wang and David Gerard
}
