% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_estdr}
\alias{hex_estdr}
\title{Wang's code to estimate double reduction rate}
\usage{
hex_estdr(NN, niter = 100)
}
\arguments{
\item{NN}{A vector of genotype counts.}
}
\description{
Wang's code to estimate double reduction rate
}
\examples{
NN <- c(29, 21, 17, 10, 10, 10, 23)

}
\author{
Jing Wang and David Gerard
}
