% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_onegen}
\alias{hex_onegen}
\title{One generation of random mating}
\usage{
hex_onegen(yww, alpha)
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{alpha}{The double reduction rate (assumed known).}
}
\value{
The next generation's genotype frequencies
}
\description{
One generation of random mating
}
\author{
Jing Wang and David Gerard

David Gerard
}
