% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_recursive}
\alias{hex_recursive}
\title{Get asymptotic frequencies from observed genotype frequencies.}
\usage{
hex_recursive(yww, niter = 8, alpha = 0)
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{niter}{The number of EM iterations.}

\item{alpha}{The double reduction rate (assumed known).}
}
\value{
The genotype frequencies at equilibrium.
}
\description{
Get asymptotic frequencies from observed genotype frequencies.
}
\examples{
yww <- c(29, 21, 17, 10, 10, 10, 23) / 120
hex_recursive(yww = yww, niter = 8, alpha = 0)

}
\references{
\itemize{
\item{J. Wang, L. Feng, S. Mu, A. Dong, J. Gan, Z. Wen, J. Meng, M. Li, R. Wu, and L. Sun. Asymptotic tests for Hardy-Weinberg equilibrium in hexaploids. \emph{Horticulture Research}, 9, 05 2022. \doi{10.1093/hr/uhac104}.}
}
}
\author{
Jing Wang and David Gerard
}
