% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_rmtest}
\alias{hex_rmtest}
\title{Test for random mating in autohexaploids}
\usage{
hex_rmtest(yww, nind, niter = 30, method = c("correct", "incorrect"))
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{nind}{The number of individuals in the sample.}

\item{niter}{The number of EM iterations.}

\item{method}{Should we calculate the chi-squared statistic the correct way
(\code{"correct"}) or the way that Wang et al calculated it
(\code{"incorrect"})?}
}
\description{
Calls \code{\link{hex_em}()} and runs a chi-squared test on the resulting
genotype frequencies.
}
\examples{
nvec <- c(29, 21, 17, 10, 10, 10, 23)
nind <- sum(nvec)
yww <- nvec / nind
hex_rmtest(yww = yww, nind = nind, method = "correct")

## reproduce 6.649 value from page 5 of Wang et al (2022)
hout <- hex_em(yww = yww, niter = 30)
rvec <- hex_recursive(yww = hout$q, niter = 8, alpha = 0)
sum((yww - rvec)^2 / rvec)

}
\references{
\itemize{
\item{J. Wang, L. Feng, S. Mu, A. Dong, J. Gan, Z. Wen, J. Meng, M. Li, R. Wu, and L. Sun. Asymptotic tests for Hardy-Weinberg equilibrium in hexaploids. \emph{Horticulture Research}, 9, 05 2022. \doi{10.1093/hr/uhac104}.}
}
}
\author{
Jing Wang and David Gerard
}
