% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexocto-package.R
\docType{package}
\name{hexocto-package}
\alias{hexocto}
\alias{hexocto-package}
\title{Code from Wang et al (2021) and Wang et al (2022)}
\description{
Takes the code from Wang et al (2021) and Wang et al (2022)
and places it in package form so that it is easier to compare.
I would not recommend using this package for real work.
You should use the \href{https://cran.r-project.org/package=hwep}{\code{hwep}}
package from Gerard (2021).
}
\details{
The original repos with the original code are
\url{https://github.com/CCBBeijing/hexaploid} and
\url{https://github.com/CCBBeijing/OctoploidDeer}.

If those repos are ever deleted or made private, you can see my forks at
\url{https://github.com/dcgerard/hexaploid} and
\url{https://github.com/dcgerard/OctoploidDeer}.
}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

Authors:
\itemize{
  \item Jing Wang
}

Other contributors:
\itemize{
  \item NSF DBI 2132247 (Supports David Gerard's work) [funder]
  \item The Forestry and Grassland Science and Technology Innovation Youth Top Talent Project of China (No. 2020132608) (Supports Jing Wang's work) [funder]
  \item The National Natural Science Foundation of China (No. 31870689) (Supports Jing Wang's work) [funder]
  \item The National Key Research and Development Program of China (2018YFD1000401) (Supports Jing Wang's work) [funder]
}

}
\keyword{internal}
