% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octo.R
\name{octo_onegen}
\alias{octo_onegen}
\title{Update genotype frequencies to next generation in octoploids}
\usage{
octo_onegen(yww, alpha)
}
\arguments{
\item{yww}{The current genotype frequencies}

\item{alpha}{The double reduction rate}
}
\value{
The next generation's genotype frequencies
}
\description{
Uses code and method of Wang et al (2021)
}
\examples{
yww <- c(0.1, 0.1, 0.15, 0.1, 0.2, 0.1, 0.05, 0.1, 0.1)
octo_onegen(yww = yww, alpha = 0)

hwep::freqnext(freq = yww, alpha = c(0, 0))

}
\references{
\itemize{
\item{J. Wang, X. Lv, L. Feng, A. Dong, D. Liang, and R. Wu. A tracing model for the evolutionary equilibrium of octoploids. \emph{Frontiers in Genetics}, 12, 2021. \doi{10.3389/fgene.2021.794907}.}
}
}
\author{
Jing Wang and David Gerard
}
