% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octo.R
\name{octo_recursive}
\alias{octo_recursive}
\title{Get asymptotic frequencies from observed genotype frequencies}
\usage{
octo_recursive(yww, niter = 20, alpha = 0)
}
\arguments{
\item{yww}{The observed genotype frequencies.}

\item{niter}{The number of EM iterations.}

\item{alpha}{The double reduction rate (assumed known).}
}
\value{
The genotype frequencies at equilibrium
}
\description{
This is from Wang et al (2021)
}
\examples{
yww <- c(0.1, 0.1, 0.15, 0.1, 0.2, 0.1, 0.05, 0.1, 0.1)
octo_recursive(yww = yww, niter = 50, alpha = 0)

## Should be binomial proportions with alpha = 0, but not
r <- sum(0:8 * yww) / 8
dbinom(x = 0:8, size = 8, prob = r)

}
\references{
\itemize{
\item{J. Wang, X. Lv, L. Feng, A. Dong, D. Liang, and R. Wu. A tracing model for the evolutionary equilibrium of octoploids. \emph{Frontiers in Genetics}, 12, 2021. \doi{10.3389/fgene.2021.794907}.}
}
}
\author{
Jing Wang and David Gerard
}
