% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_estdr}
\alias{hex_estdr}
\title{Wang's code to estimate double reduction rate}
\usage{
hex_estdr(NN, niter = 100, tol = 10^-4)
}
\arguments{
\item{NN}{A vector of genotype counts.}

\item{niter}{The number of iterations in the EM algorithm.}

\item{tol}{The stopping criterion for the EM algorithm.}
}
\description{
Wang's code to estimate double reduction rate
}
\examples{
## Generate data under assumed mode
alpha <- 0.2
p <- c(0.3, 0.2, 0.35, 0.15)
p <- p / sum(p)
q <- stats::convolve(p, rev(p), type = "open")
q1 <- hex_onegen(q, alpha = alpha)
NN <- c(stats::rmultinom(n = 1, size = 100, prob = q1))
hex_estdr(NN = NN)
p

}
\author{
Jing Wang and David Gerard
}
