% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexocto-package.R
\docType{package}
\name{hexocto-package}
\alias{hexocto}
\alias{hexocto-package}
\title{Code from Wang et al (2021) and Wang et al (2022)}
\description{
Takes the code from Wang et al (2021) and Wang et al (2022)
and places it in package form so that it is easier to compare.
I would not recommend using this package for real work.
You should use the \href{https://cran.r-project.org/package=hwep}{\code{hwep}}
package from Gerard (2021).
}
\details{
The original repos with the original code are
\url{https://github.com/CCBBeijing/hexaploid} and
\url{https://github.com/CCBBeijing/OctoploidDeer}.

If those repos are ever deleted or made private, you can see my forks at
\url{https://github.com/dcgerard/hexaploid} and
\url{https://github.com/dcgerard/OctoploidDeer}.
}
\references{
\itemize{
\item{D. Gerard. Double reduction estimation and equilibrium tests in natural autopolyploid populations. \emph{bioRxiv}, 2021. \doi{10.1101/2021.09.24.461731}.}
\item{J. Wang, L. Feng, S. Mu, A. Dong, J. Gan, Z. Wen, J. Meng, M. Li, R. Wu, and L. Sun. Asymptotic tests for Hardy-Weinberg equilibrium in hexaploids. \emph{Horticulture Research}, 9, 05 2022. \doi{10.1093/hr/uhac104}.}
\item{J. Wang, X. Lv, L. Feng, A. Dong, D. Liang, and R. Wu. A tracing model for the evolutionary equilibrium of octoploids. \emph{Frontiers in Genetics}, 12, 2021. \doi{10.3389/fgene.2021.794907}.}
}
}
\author{
\strong{Maintainer}: David Gerard \email{gerard.1787@gmail.com} (\href{https://orcid.org/0000-0001-9450-5023}{ORCID})

Authors:
\itemize{
  \item Jing Wang
}

Other contributors:
\itemize{
  \item NSF DBI 2132247 (Supports David Gerard's work) [funder]
  \item The Forestry and Grassland Science and Technology Innovation Youth Top Talent Project of China (No. 2020132608) (Supports Jing Wang's work) [funder]
  \item The National Natural Science Foundation of China (No. 31870689) (Supports Jing Wang's work) [funder]
  \item The National Key Research and Development Program of China (2018YFD1000401) (Supports Jing Wang's work) [funder]
}

}
\keyword{internal}
